requirements_frozen.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
attrs==24.2.0
cfgv==3.4.0
coloredlogs==15.0.1
distlib==0.3.8
filelock==3.15.4
fire==0.6.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
grpcio==1.66.1
humanfriendly==10.0
hydra-core==1.3.2
identify==2.6.0
idna==3.8
iniconfig==2.0.0
Jinja2==3.1.4
lightning==2.4.0
lightning-utilities==0.11.6
Markdown==3.7
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
nodeenv==1.9.1
numpy==2.1.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.19.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
-e git+ssh://git@github.com/lfagotbouquet/template_python.git@33565cfe21e91da1051f647d43301e108850a3e3#egg=PROJECT_NAME
protobuf==5.28.0
pytest==8.3.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2024.1
PyYAML==6.0.2
six==1.16.0
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
torch==2.4.0
torchmetrics==1.4.1
torchvision==0.19.0
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.3
Werkzeug==3.0.4
yarl==1.9.6
